moved emptyPage in the folder for shared components
