+++ src/App.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import AdminPage from "./components/AdminPage";
 import CheckoutPage from "./components/CheckoutPage";
 import DetailPage from "./components/DetailPage";
-import EmptyPage from "./components/EmptyPage";
+import EmptyPage from "./components/shared/EmptyPage";
 import Layout from "./components/Layout";
 import NotFound from "./components/NotFound";
 import ProductPage from "./components/ProductPage";
